# Use heuristic to determine voro++ block sizes based on particle density.