--- a/ChangeLog.md
+++ b/ChangeLog.md
@@ -11,6 +11,7 @@ and this project adheres to
 
 ### Changed
 * The make\_random\_box system method no longer overwrites the NumPy global random number generator state.
+* The Voronoi class uses smarter heuristics for its voro++ block sizes, resulting in significant performance gains for large systems.
 
 ### Fixed
 * The from\_box method correctly passes user provided dimensions to from\_matrix it if is called.