--- a/cpp/locality/Voronoi.cc
+++ b/cpp/locality/Voronoi.cc
@@ -34,13 +34,20 @@ void Voronoi::compute(const freud::locality::NeighborQuery* nq)
     {
         boxLatticeVectors[2] = box.getLatticeVector(2);
     }
-    // TODO: This container uses 3 blocks in x, y, and z, and an initial
-    // memory allocation of 3, which should be improved. Ideally, this code
-    // should use a pre_container or implement its own heuristics to choose
-    // a number of blocks.
+
+    // This heuristic for choosing blocks is based on the voro::pre_container
+    // guess_optimal method. By computing the heuristic directly, we avoid
+    // having to create a pre_container. This saves time because the
+    // pre_container cannot be used to set up container_periodic (only
+    // non-periodic containers are compatible).
+    float block_scale = std::pow(n_points / (voro::optimal_particles * box.getVolume()), 1.0/3.0);
+    int voro_blocks_x = int(box.getLx() * block_scale + 1);
+    int voro_blocks_y = int(box.getLy() * block_scale + 1);
+    int voro_blocks_z = int(box.getLz() * block_scale + 1);
+
     voro::container_periodic container(boxLatticeVectors[0].x, boxLatticeVectors[1].x, boxLatticeVectors[1].y,
                                        boxLatticeVectors[2].x, boxLatticeVectors[2].y, boxLatticeVectors[2].z,
-                                       3, 3, 3, 3);
+                                       voro_blocks_x, voro_blocks_y, voro_blocks_z, 3);
 
     for (size_t query_point_id = 0; query_point_id < n_points; query_point_id++)
     {