## Description
I was investigating a large system (500,000 particles) and the performance of Voronoi slowed to a crawl. When I implemented voro++, I hard-coded the block sizes and it seemed to perform fine for the test systems. However, those hard-coded sizes were terrible for performance with this large system. I have fixed the TODO item in the code, by implementing a heuristic that aims for an average of 5.6 particles per block (which is the optimal value provided by voro++).

## Motivation and Context
Significant performance improvement for Voronoi in large systems (>10k particles).

## How Has This Been Tested?
Existing tests pass.

Performance test:
```python
import freud
from datetime import datetime

voro = freud.locality.Voronoi()
for N in 500, 5000, 50000, 500000:
    system = freud.data.make_random_system(10, N, seed=42)
    start = datetime.now()
    voro.compute(system)
    end = datetime.now()
    print(N, 'particles', end-start, 'elapsed')
```

Results:

Particles | With Heuristics | master | Speedup
-- | -- | -- | --
500 | 0.020 | 0.024 | 1.2x
5000 | 0.181 | 0.523 | 2.9x
50000 | 1.679 | 16.893 | 10.1x
500000 | 14.410 | 1114.164 | 77.3x

## Types of changes
<!-- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds or improves functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist:
<!-- Put an `x` in all the boxes that apply. If you're unsure about any of
     these, don't hesitate to ask. We're here to help! -->
- [x] I have read the [**CONTRIBUTING**](https://github.com/glotzerlab/freud/blob/master/CONTRIBUTING.md) document.
- [x] My code follows the code style of this project.
- [x] I have updated the documentation (if relevant).
- [x] I have added tests that cover my changes (if relevant).
- [x] All new and existing tests passed.
- [x] I have updated the [credits](https://github.com/glotzerlab/freud/blob/master/doc/source/reference/credits.rst).
- [x] I have updated the [Changelog](https://github.com/glotzerlab/freud/blob/master/ChangeLog.md).